+++ source/extensions/omni.isaac.benchmarks/omni/isaac/benchmarks/tests/test_benchmark_ros_camera.py
@@ -8,15 +8,13 @@
 #
 
 
+import numpy as np
 import omni.kit.test
 from omni.isaac.core.utils.rotations import euler_angles_to_quat
-
-# from omni.isaac.core.prims._impl.single_prim_wrapper import set_default_state
-from omni.isaac.core.utils.viewports import destroy_all_viewports, get_viewport_names
-from pxr import Gf, UsdGeom
+from omni.isaac.sensor import Camera
 
 from ..utils.base_isaac_benchmark import BaseIsaacBenchmark
-from ..utils.helper import add_ros_camera
+from ..utils.helper import add_ros1_camera
 
 TEST_NUM_APP_UPDATES = 60 * 10
 
@@ -24,12 +22,18 @@ TEST_NUM_APP_UPDATES = 60 * 10
 class TestBenchmarkRos1Camera(BaseIsaacBenchmark):
     async def setUp(self):
         await super().setUp()
-        destroy_all_viewports(destroy_main_viewport=False)
+        from omni.isaac.ros_bridge.scripts.roscore import Roscore
+        from omni.isaac.ros_bridge.tests.common import wait_for_rosmaster
+
+        self._roscore = Roscore()
+        await wait_for_rosmaster()
+
         pass
 
     async def tearDown(self):
         await super().tearDown()
-        destroy_all_viewports(destroy_main_viewport=False)
+        self._roscore.shutdown()
+        self._roscore = None
         pass
 
     # ----------------------------------------------------------------------
@@ -45,61 +49,38 @@ class TestBenchmarkRos1Camera(BaseIsaacBenchmark):
         timeline = omni.timeline.get_timeline_interface()
         timeline.play()
 
+        cameras = []
+
         for i in range(n_camera):
-
-            # add a camera on stage
-            camera_path = "/Cameras/Camera_" + str(i)
-            if i == 0:
-                viewport_name = "Viewport"
-            else:
-                viewport_name = "Viewport " + str(i)
-            stage = omni.usd.get_context().get_stage()
-            camera_prim = stage.DefinePrim(camera_path, "Camera")
-            camera_translation = Gf.Vec3f([-8, 13, 2.0])  # these positions are used for full_warehouse.usd
-            if "xformOp:translate" not in camera_prim.GetPropertyNames():
-                UsdGeom.Xformable(camera_prim).AddTranslateOp()
-            camera_prim.GetAttribute("xformOp:translate").Set(camera_translation)
-            q = euler_angles_to_quat([90, 0, 90 + i * 360 / n_camera], degrees=True)
-            camera_orientation = Gf.Quatf(q[0], q[1], q[2], q[3])
-            if "xformOp:orient" not in camera_prim.GetPropertyNames():
-                UsdGeom.Xformable(camera_prim).AddOrientOp()
-            camera_prim.GetAttribute("xformOp:orient").Set(
-                camera_orientation
-            )  # rotate cameras to look at slightly different view, repeated views affect fps
-
-            # add corresponding ROS camera publisher (that also creates viewports)
-            graph_path = "/ROS/ROS_camera_" + str(i)
-            ros_topic = "/rgb_" + str(i)
-            add_ros_camera(
-                camera_prim_path=camera_path,
-                graph_path=graph_path,
-                camera_topic=ros_topic,
-                viewport_name=viewport_name,
-                viewport_resolution=resolution,
+            cameras.append(
+                Camera(
+                    prim_path="/Cameras/Camera_" + str(i),
+                    position=np.array([-8, 13, 2.0]),
+                    resolution=resolution,
+                    orientation=euler_angles_to_quat([90, i * 360 / n_camera, 0], degrees=True),
+                )
             )
+
+            await omni.kit.app.get_app().next_update_async()
+            cameras[i].initialize()
+
+            rp_path = cameras[i].get_render_product_path()
+            add_ros1_camera(rp_path, "/Graphs/Camera_" + str(i), "/rgb_" + str(i), "sim_camera_" + str(i))
             await omni.kit.app.get_app().next_update_async()
 
-            while viewport_name not in get_viewport_names():
-                await omni.kit.app.get_app().next_update_async()
-            if i > 0:
-                main_viewport = omni.ui.Workspace.get_window("Viewport")
-                new_viewport = omni.ui.Workspace.get_window(viewport_name)
-                new_viewport.dock_in(main_viewport, omni.ui.DockPosition.RIGHT, 1.0 / n_camera)
+        self.stop_runtime()
+        await self.store_measurements()
+
+        self.set_phase("benchmark")
+        self.start_collecting_frametime()
+
+        while self.get_num_frames() < TEST_NUM_APP_UPDATES:
             await omni.kit.app.get_app().next_update_async()
 
-            self.stop_runtime()
-            await self.store_measurements()
+        self.stop_collecting_frametime()
+        await self.store_measurements()
 
-            self.set_phase("benchmark")
-            self.start_collecting_frametime()
-
-            while self.get_num_frames() < TEST_NUM_APP_UPDATES:
-                await omni.kit.app.get_app().next_update_async()
-
-            self.stop_collecting_frametime()
-            await self.store_measurements()
-
-            timeline.stop()
+        timeline.stop()
 
     async def test_benchmark_ros1_1_camera_720p(self):
         await self.benchmark_ros1_camera(1, [1280, 720])