+++ source/extensions/omni.isaac.range_sensor/python/scripts/samples/generic_info.py
@@ -199,8 +199,6 @@ class Extension(omni.ext.IExt):
             # we want to make sure we can see the sensor we made, so we set the camera position and look target
             set_camera_view(eye=[-5.00, 5.00, 5.00], target=[0.0, 0.0, 0.0], camera_prim_path="/OmniverseKit_Persp")
 
-            #
-            # self._editor_event_subscription = self._editor.subscribe_to_update_events(self._on_editor_step)
             self._editor_event_subscription = (
                 omni.kit.app.get_app().get_update_event_stream().create_subscription_to_pop(self._on_editor_step)
             )